Fix uninstall when modules are present
When we were attempting to perform uninstallation it was resulting in a
fatal error if any modules were installed, because we weren’t loading
all of the necessary dependencies for the module uninstallation process.

We missed this because our tests run without any modules installed.
This is now changed.

Fixes #380