Un/installer: support wildcard meta keys
Based on the wildcard support for options.

While we were creating the tests for this, we realized that one of our
existing tests was actually broken because it was passing the
parameters of `get_user_option()` in the wrong order. `$user_id` was
expected to come first, but it is actually second. Anyway, we’ve now
fixed this, which means the tests logic had to be changed just slightly.

See #395