Pass hook mode to reaction store constructor
To do this properly, we have updated the children classes registry to
automatically pass the parent slug in addition to the child slug (since
the mode is the parent slug), because otherwise this becomes
problematic, because otherwise all of the classes would be constructed
with the same mode.

This change also revealed some old cruft which we have cleaned up. It
also necessitated a change of the signature of the hook arg constructor.

I opted not to add a `get_mode()` method to the reaction store class
right now, because I’m not sure whether this should be a part of the
interface or not. Requires more thought.

Fixes #411