Deprecate *_network_option() functions
In favor of the new `*_maybe_network_option()` functions. Also
deprecate the `network` context of the get array option function. We
don’t introduce a `maybe_network` context there, because we have a get
maybe network array option function which should be used instead. It
makes sense to keep that function around, because it allows you to pass
the `$maybe_network` arg, which can’t be done using the old array
function. And there’s no reason to make a duplicate method of doing
this, so that’s why we haven’t added the `maybe_network` context to the
old array function.

Fixes #305