Remove l10n files that don't need to be bundled
These are either complete (Chinese) or are too incomplete to be worth
bundling anymore. When they are completed they’ll be sent out to users
as language packs. Many of these have never been release with
WordPoints before anyway. We’ve only kept a few translations that have
been released before and are somewhat complete.

See #355