Make most admin messages dismissible
Also turn some errors into warnings, and success messages into info.

The only ones that I’ve decided not to make dismissible are those that
will be shown again if the screen is reloaded anyway, i.e., they aren’t
related to an action that the user has just performed. If we want to
make these dismissible, we’ll want to associate an option with them or
something.

See #222