Add action to delete points type function
We also move the actual deletion of the points type’s settings to the
end of the function. We don’t bother to check the result of that
anymore either, since it alone doesn’t indicate failure. (We don’t
check the result of every other operation in the function, FWIW.)

Fixes #361