Different message when adding points type
When editing a points types settings in the per-site admin, we show a
message to the user when WordPoints is network active, warning them
that all sites across the network will be affected. This same message
was also being displayed when a new points type was being added, but
now we’ve added a different message for that case.

Fixes #415