Focus on the first input of a new condition
When the condition is added, we focus on the first visible input (after
the delete button).

I’ve overridden the `isNew()` method for the `Condition` model, so that
it now checks whether the condition is present in the reaction it is
associated with. This is necessary because conditions always have an ID
from the moment that they are created.

Fixes #416