Refactor autoloader to use class-maps
This simplifies the autoloader, improves performance, but slightly
increases memory consumption.

We’ve also updated the Grunt configuration to use the updated
autoloader task included in the WordPoints dev-lib, so that the class
maps are auto generated. This update also automatically creates the
index files for all class directories, not just `src/includes/classes`.

Fixes #421