Better compat for imported post publish hooks
Introduces a legacy version of the post publish event that only
reverses when a post is deleted. This event is only registered by the
points component, and only when legacy support is enabled by an option
flag in the database. In order to support this, the post delete action
for each post type is now registered by default. Previously it was only
registered for the attachments, because `media_upload` is the only
event that uses it by default.

Because these legacy events are only registered after support is
enabled, it is necessary to manually loop the function over the list of
post types (like with `array_map()`) before importing. Otherwise the
reactions wouldn’t be created for that event, because validation of the
event would fail if it wasn’t registered.

The new repeat blocker extension is also utilized to ensure that the
reactions won’t fire when a post is published, if points had already
been awarded for that post before importing it to the new hooks API.

The PHPUnit autoloader function has been updated for easier testing of
the new event.

Fixes #388