Use supplied log meta keys in legacy extensions
The legacy reversals and repeat blocker hook extensions included with
the points component query the points logs in order to maintain correct
behavior for imported points hooks. The queries involve matching on the
ID of the primary entity for that hook, by querying the points log
meta. The meta key to query on was previously being guessed by these
extensions automatically, but the guesses were actually incorrect.
Instead of assuming that every imported points hook will follow a
single pattern, we’ve changed so that the meta key to use is supplied
in the reaction metadata. For some hooks no metadata is added, and so
there is no meta key to query on. (This is done because the primary arg
is the user that the transaction is targeting already, so saving that
as metadata was considered redundant in the old points hooks.)

We’ve updated the tests for each of these extensions accordingly, and
included tests that actually utilize the old points hooks, so that we
can be sure that everything really does integrate properly. These tests
would have failed previously.

We also update the importer to support providing the legacy meta key
when importing hooks. And we’ve updated the importer tests to utilize
this feature. I’ve also added a check to one of the tests to verify
that the hook is actually imported.

Fixes #435