Use GMT dates for hook hit logs
And just about everywhere else as well.

One very significant update required because of this is the periods
hook extension. It needs to calculate the absolute periods relative to
the site’s GMT offset, and so we have to update it to do that properly.
(This is really only significant with period lengths of more than an
hour.) The unit tests have also been improved to test this more
thoroughly.  The offsets of 5 and -5 used in the tests weren’t chosen
for any particular reason.

I’ve also updated the logic for the relative periods, because it was
confusing the way the math was done before. So I flipped it.

We stop using `gmdate()` as it is not necessary, and so using it can be
confusing.

See #439
Fixes #440