PHPUnit: only load autoloader if not exists
When modules are running their tests using the dev lib bootstrap, the
dev lib version included with the module will load the autoloader
already. So when WordPoints’s bootstrap is included, it will already
exist. Trying to load the class again from WordPoints’s copy of the dev
lib will result in a fatal error.

See #499