Convert UI PHPUnit tests to Codeception tests
We are able to deprecate some of the code which related to these tests,
as well as removing the tests themselves. We don’t remove the helper
code just in case something else is using it. So we don’t stop
including the util-getopt class, but we do stop constructing it. We
cannot use `_deprecated_function()` in all locations, since some of the
functions are expected to be called before WordPress is loaded.

Fixes #373