Fix fatal error in new PHPUnit points log factory
The factory collection that it is constructed with only contains the
WordPoints factories, but we were assuming that it was the core
factories (since this was the case with the old implementation).

We have to make the `factory()` method public in our child test case so
that we can retrieve the core factory collection, since that method is
protected on the parent.

See #474