Use WPPPB for PHPUnit tests
In the process, we switch to autoloading the test dependencies. One
side-effect of this is that we found an issue in the filesystem mock
package, and so had to fix that and update to the fixed version.

The uninstall tester tools have been merged into WPPPB, and slightly
modified. I will no longer be maintaining them separately, so we remove
them as a dependency, and update our uninstall tests accordingly. Among
the changes there is that WPPPB now assumes that the uninstall tests
will be specified in a separate config file, and so we move them there.
The uninstall test case has been updated to reflect the fact that the
plugin file name, network-activation status, and simulation file, are
automatically detected by WPPPB.

As our extension of WPPPB now takes care of loading the plugin and
components, we no longer need to load them manually. Consequently, the
functions that did that have been deprecated.

See WordPoints/dev-lib#161