Fix module deletion UI
Deleting a module through the admin screens was broken, resulting in PHP notices being thrown, module information not being displayed on the verification screen, and the module not actually being deleted. All of this is now fixed with just two little changes.

Oh, and did I mention that we added Codeception tests to prove it?

Fixes #509