Create ConditionGroup views on render
Rather than beforehand. Creating the views early causes views to be created which are never rendered, but which are listening for events coming from the reaction. This will result in errors when the view assumes that it has already been rendered in those event callbacks. Instead, we wait to create the view until just before we are about to render it.

Fixes #523