Fix add condition form after cancel
After cancelling editing a reaction, the add condition form for that reaction would be not show the arg and condition selectors, only the add and cancel buttons. This is now fixed, by re-attaching the views for the selectors to the condition groups' view.

The jQuery selectors for the selector views have also been corrected to be more specific, whereas before they would also target sub-condition groups' views as well.

An acceptance test has been included, and the other tests have been updated to match teh new behavior of `addCondition()` no longer calling `edit()` directly. (The dev-lib will be updated in a later commit.)

Fixes #520