Fix cancel/close/save buttons on cancel
When a reaction is open but unchanged, it should show a Close and (disabled) Save buttons. When it is changed, we hide the Close button, enable the Save button, and show a Cancel button. Hitting the Cancel button will clear any changes, and should cause the button to be hidden again, the Save button disabled, and the Close button shown.

This has always been the intention, but recent changes in #517 have cause the Cancel button to be re-shown and teh Close button hidden after it is initial clicked and the inverse performed. The Save button was also not being disabled as it should have been. All of this is now fixed, and the behavior added to the cancel update reaction Codeception acceptance test.

Fixes #524