Don't add duplicate Rate Limits on cancel
When cancelling editing a reaction, the `render:fields` event is triggered. In the Periods extension we hook into that to render the Rate Limits section and append it to the reaction fields view. However, on cancel, we need to replace the already existing Rate Limits view instead of appending a new one.

Fixes #516