Entities API: accept entity GUIDs
Previously `Entity::set_the_value()` would accept an entity ID or entity object, and always defaulted to the current context. Now it is possible to pass in the full GUID for an entity, which will properly set the context and switch to that context momentarily in order to properly retrieve the entity object.

Fixes #535