Multisite switched state class
Extracts the code for tracking the switched state on multisite out of
the un/installer class and into its own class that can be used
elsewhere in the code. We will use it with the points logs momentarily.
This allows us to switch without ever having to restore until we are
done, saving us from switching back all of the time.

Currently the class doesn’t support saving multiple snapshots, just
one. If we need that feature in the future, and proper snapshotting
hasn’t been introduced in core, we can add it here.

In the process of running the unit tests for the new class I realized
that creating the sites adds a bit of extra time to each test, since
the tables for each site have to be created. WordPress core has been
moving toward sharing some fixtures like this between all of the tests
in a testcase. I decided to do the same, although I created our own
method of doing so that will also automatically clean up after
ourselves in the parent testcase, all that the child testcase has to do
is ask for the fixtures to be created, and the parent testcase will
take care of everything else. Currently no special attributes to give
the fixtures can be specified, but we’ll expand on this in the future.
In order to gain in cleaning up afterward, I’ve introduced a new
interface that our factories can implement so that they can provide a
delete method in addition to the create an update ones. We don’t
implement it anywhere yet, but we’ll do that in the future when we need
to actually use it. For the core object types, we’ll have to add
methods to the parent testcase instead. For some of these that we
already have in child classes, I’ve pulled them up.

Note that shared fixtures have to be used with care, and generally only
in tests where the fixtures are only read, not modified in any way.

See #536, #424