Deep registry: require array hierarchy
The hierarchy of parent slugs was previously allowed to be a string
when the class in question was a child of a single parent, instead of a
grandchild. This was not entirely standardized internally, which could
allow for the classes to be constructed with the slugs in either
format, depending on how they were passed in. So it was decided that it
would be  best to standardize on always requiring arrays to be passed
in, so that it would be obvious that classes would always be
constructed with arrays (which they now will, of course).

See #541