Deep registry: add pass_slugs setting
This can be used to disable the passing of the class slugs to the
constructors when the objects are created. Currently this setting can
be manipulated via the `update_setting()` method. However, in the
future we will probably change things so that this setting can be
passed in to the constructor, and then that method may be deprecated.

The helper function located on the base registry also had to be
updated. There is some duplication of logic there, for optimization.

See #541