Add slug-less multilevel registry
Instead of introducing setter methods for `pass_slugs` setting, we just
set it in the child “slugless” class. This causes the registry to
always be slugless from the time it is constructed, rather than it
being set that way at a later point. Doing otherwise was causing issues
when mocking the apps in the tests, because apps using slugless
registries were setting that feature when the registry initialization
action was called, which we don’t call when the tests are mocking the
apps.

See #541