Entity restrictions API
Introduces a new sub-app of the entities app that provides a
full-fledged entity restriction API. The benefit of this new API over
the old simple function is that it allows you to check not just if a
user can view an entity but also whether the restriction applies to the
particular entity in question at all—meaning that you can check if some
other users might not be able to view the entity even though the user
in question can.

This deprecates the old interface that was implemented by the classes
for entities with restricted visibility, as well as the filter that was
in the old function. The logic from the function has been moved to a
legacy restriction object, and the function itself has been updated to
use the new API, and has not been deprecated.

The entities that previously implemented the deprecated interface do
not anymore, the `user_can_view()` methods have been retained but
deprecated, and utilize the new API. The tests have been updated
accordingly, and one part of this required us to use real slugs for the
entities, that are actually registered, since otherwise the
“unregistered” restriction would cause the checks to always return
false.

The only place that the old function was currently in use was the
points logs, and they will be updated to use the new API in a later
commit, along with their own restrictions API.

Fixes #541