Don't mock base context methods
In these tests we are supposed to be testing the actually parent method
logic, no the mock’s logic, so we need to not mock those methods. We
originally did this but then ended up expanding the mock and adding
these methods later, without realizing that it would break these tests.

Fixes #479