Points logs viewing restrictions API
Introduces an API for restricting points logs from being viewed by certain users. The app is a children class registry, where the parent slugs are the log types. Generic restrictions that apply to all types of logs can also be registered with `'all'` as the parent slug. This registry constructs the restrictions without the slugs, since the log type can be determined from the log itself, which is passed to the restrictions' constructors, and the slug of the particular class for that log type is not needed. The restrictions are not mean to be retrieved directly, but via the `get_restrictions()` method, which will automatically include the generic restrictions along with any restrictions that apply specifically to that log type.

The value of the new API is that it allows one to not only determine whether a particular user can view a log, but whether there are any restrictions that apply to that particular log at all, and if so, to get a description of each of them.

The legacy function and the filters are retained, but the filters are deprecated. The old points hooks will be updated to use the new API shortly.

The points logs view is updated to use the new API directly, so that it will also be able to get the description of the applicable restrictions, which it will use shortly.

The fixture handling in the testcase has been updated extensively, with the additional features documented inline. This has also resulted in some updates within the points logs factory.

See #536, #424