Fix PHPUnit tests for restrictions API
The introduction of the points logs restrictions API involved several modifications of the PHPUnit tests, and there were a few remaining issues that we didn't spot.

First, there were some errors being caused by us assuming that `$extra_fixture_ids` would always being an array, even though we set it to `null` after the fixtures are destroyed. I've decided to just reset these values to empty arrays, instead of `null`, so we won't have these kind of issues.

We also realized that after one testcase had run our fixtures were no longer being saved when we created them. This is because the parent `setUp()` method sets `autocommit` to `0` when it starts the transactions, meaning anything that isn't explicitly committed will be discarded when a new transaction is created. This didn't reveal itself as a problem for our fixtures when we were running a single testcase, because we call `parent::setUp()` *after* we create the fixtures, so that when it first runs no transactions have been started and autocommitting hasn't yet been disabled. But after one testcase has run, autocomitting has then been disabled, and so it is necessary for us to explicitly commit our fixtures or else they will be discarded when we call `parent::setUp()` and a new transaction is started.

We also discovered that there were some errors in tests that were expecting a points type to exist, as created by `$this->create_points_type()`. It turned out that that method wasn't creating a points type after we'd created some by our fixture code (as we do for points logs). This was because when we delete the points type created with the fixtures, the option in the database will remain as an empty array. And because we create the points type using `add_option()`, this would cause it not to be created since the option would already exist. I've just updated the code to use `update_option()` instead, which should cause no issues, since there is little fear of overwriting anything since anywhere this is being called it is expected to be creating something where nothing exists anyway, and wouldn't be working if the option did.

Finally, there were a few race-condition issues in the points logs view tests, and update them to expect the correct descending ordering.

See #536, #544