Match interface and implementation signatures
In PHP 5.2, in order for a method argument to have a default value in an implementation, the interface must also specify that same default value. In later versions of PHP this is no longer required, so we didn't notice until now that our multilevel registry was providing a default value for `$parent_slugs` where appropriate, but the interface wasn't.

See #541