Use entity GUIDs in hooks restrictions
The points logs viewing restrictions for the points hooks was only passing in the ID of the entities, but greater precision can be obtained by passing in the GUID of the entity instead. To this end, we update the Points reactor to save the GUID for each entity, as well as the ID. Then in the points logs vewing restriction for the hooks API we utilize the GUIDs when they are available, but fall back to just the ID for back-compat.

The tests for the hooks restriction is also updated to use more fixtures. This prompted us to fix a bug in the base testcase's fixture handling, that was causing an error about an array being provided where a string was expected, because the points logs meta arg is an array of values.

See #424, #536, #535, #479