Don't refire hooks on post update
The post publish and post depublish actions were both firing when a post was updated, because the `transition_post_status` action fires even when both the old and new statuses are the same. Since these two hooks were just listening for the new and old statuses to be `publish`, respectively, when updating a post and both the old and new statuses were `publish`, both of these actions would fire, resulting in points being awarded and removed at once.

Note that our tests for these hooks were actually incorrect, as they were expecting the log to be reversed by setting the post status to `publish` instead of `draft`—meaning that our tests were actually relying on this behavior!

Fixes #550