Use a cryptographically secure PRNG
This random string is used to generate an nonce for when we are checking module compatibility during a breaking update of the plugin. Even though a lack of cryptographic security is not really exploitable here, we'll follow best practice and use it anyway. There isn't much incentive for exploitation either, since the most that an attacker could do is view a list of installed modules. And they would have only a window of a few seconds in which to try and exploit the issue, when the site was performing a breaking update of WordPoints, which is a rare occurrence anyway.

Thank you to [@e3amn2l](https://hackerone.com/e3amn2l) for responsibly disclosing this issue.