HTTPS all the things
Thanks to [@e3amn2l](https://hackerone.com/e3amn2l) for pointing out that we could update many of these URLs to be HTTPS, and decrease the likelyhood of MITM attacks when people click links in their admin that lead to WordPoints.org or WordPress.org. We just updated everything to be HTTPS, if the domain supported it.