Autoload core classes where possible
Adds classmaps to load core classes that are in files separate from any other code. The two shortcode classes are grouped together in a single file, but this can be safely loaded via the classmap as well.

Classes that are in files along with other code have not been included. Each of the component's classes will be handled in separate commits.

See #372