Add classmap for autoloadable points classes
Use autoloading for all classes that are in files separate from other code. The shortcode classes are all grouped together in a single file, but this is OK. We have moved the shortcode registration out of that file so that we can use autoloading for them.

Other classes that are in files mixed with other code will be handled later.

See #372