Add classmaps for autoloadable ranks classes
Introduces autoloading for all classes that are in files separate from any other code. The shortcode classes are grouped in a file together, but that is OK, we just have to move the shortcode registration into the `filters.php` file.

We can also deprecate the `rank-types.php` file, since it is no longer needed. We had kept it around before in case we introduced more rank types in the future, so that there would be a single file that would load all of them. But now that we'll be using autoloading, that will no longer be necessary.

See #372