Fix network-activating multiple modules
When network-activating a second module, the previously activated one would be accidentally deactivated.

In the process of writing the unit tests for this, we also discovered another bug in the module activation function, when attempting to activate a non-existent module when WordPoints was network active. This was fixed by moving up the module validation code.

The Codeception tests for this have to run with WordPoints network-active, we requires us to update the dev-lib.

Fixes #562