Change primary_arg_guid column name
It can now contain the GUIDs for multiple signature args, so it should be named `signature_arg_guids` to match the rest of the code. Backward compatibility is provided in the query class for the hook hits table.

We bump the WordPoints version so that the update will run on sites running the alpha version.

In the process we've also updated some test cases in the dev-lib, so the Travis config is updated to run against WordPress 4.7.

See #594