Browserify: include sourcemaps
In the unminified versions of the JS files, since these will usually only be used in debugging anyway.

We've wanted to do this for some time, but needed to wait until we were minifying the files so that the sourcemaps wouldn't be adding extra bytes to the size of the files normally used in production.

See #58