Move old points shortcode classes back
By placing them back in their old file again, they can be autoloaded only as needed. This is beneficial in this case, because they extend the new shortcode classes, and so if they are loaded every time the new classes will be as well. This way all of them will be autoloaded JIT.

See #498