Move Points Logs widget class to its own file
This class was not previously autoloaded, but now it is. This makes little difference presently, since widgets are always loaded by WordPress. However, autoloading helps us by making it possible to more easily have other classes that depend on it be autoloaded. (Which again, makes little difference, but is just good in general for consistency.)

See #498