Improve points logs widget display
The table can quickly become cramped on small sidebars, especially the description column. To alleviate this, we give the user some more options in regard to how the table is displayed:

- We allow the user to hide any of the columns of the table via the widget settings.
- We allow horizontal scrolling to be forced an teh table width to be set at 500px.

In addition, we make sure that the table does not used fixed layout, since that isn't a good experience for the nature of this table. And, we've intentionally done all of this without any inline styles.

See #333