User Points widget: extend logs widget
This allows us to easily take advantage of the new options added to the logs widget to improve display on skinny sidebars.

The `widget_body()` method of the points logs widget is now more configurable via class properties. And the `form()` method is broken up into several different functions for each set of fields, for easier sub-classing.

In the process a bug in the saving of the horizontal scrolling and hide user names options was discovered, that prevented the disabling of these options, and was fixed.

See #333