Prevent widget errors in customizer
When a widget is first added to the customizer, the settings are completely empty as it is first displayed. The result of this is that an error is shown for our points widgets, because they expect the points type to be set. To work around this, when the points log is completely empty, we now specify a default value, if possible.

Fixes #618