Blocker extension: Support cross-action type settings
Previously each action type had to be blocked separately, now blocking all action types is supported. This is needed to allow us to use this class as the disable extension.

Note that we use `wp_validate_boolean()` because values of `"true"` and `"false"` naturally result from the way that the JS hooks API sends boolean values back to the server.

In the process we also discovered some issues in the tests for the blocker extension, which was causing it to not be really tested at all in some cases (because the settings were being treated as the whole settings for the reaction, instead of just the settings for the extension).

See #470