Dynamic entity relationship: use parent slug
Previously we got the dynamic value from the relationship slug itself, but there is no reason why we can't get it from the parent slug instead, and this is more useful to us at present. Specifically, we'd like to use this so that we can introduce the `'parent'` relationship between a comment and its parent comment without making the relationship slug itself dynamic.

See #529