Fix tests for Comment Leave event
The tests for the event for hierarchical post types were failing, because we were not creating a parent post.

While we are at it, we add the parent post author to the list of expected targets, so that the tests will fail if for some reason they stop testing against the parent post author in the future.

See #529