Fix legacy Post Publish event tests
The tests for the event for hierarchical post types were failing, because we were not creating a parent post. We had already fixed this for the tests for the regular post publish event, but not the legacy version of the event registered for the points component for imported hooks.

See #422