Block updates when PHP requirements aren't met
Currently WordPoints's PHP requirements are identical to those of WordPress. Ideally, it will always be that way. However, it is possible that we will have a good reason to diverge from WordPress's requirements in the future, and in that case we want to be prepared to to so gracefully.

To that end, we've done the following here:

- Added a function to check the description of each WordPoints update to see if a particular PHP version requirement is specified.
- If an update of the plugin does have a requirement, we also provide a function that will check if it met by the current install.
- If an update's required PHP version isn't met, then we remove the update row in the plugins list table and replace it with an error notice explaining the situation and a link to more information.
- We also hide the update on the Dashboard » Updates screen.
- And we prevent the user from being able to update via the "View Details" modal.

Fixes #623