Register legacy modules screen differently
The way we were doing it before, the list table was not working properly
(for some reason the columns weren't being registered, so none of the
columns were being displayed). Rather than spend more time trying to get
to the root cause of this, we'll just use a different method of
registering the screen, which does not suffer from these problems.

See #668