legacy extensions dir filter: Check if version is set
Usually it always will be, but a sanity check seems more sane. One case
where there *will* be issues is when uninstalling after WordPoints was
installed network-wide. In that case, the version check will be invoked,
but we will not have pulled the version info from a site option, but
just a regular option. This is because `is_wordpoints_network_active()`
always returns false when WordPoints isn't active—which it isn't when it
is being uninstalled. (See #399)

See #668