Updates: use DB version when from version not specified
We allow the version to update from to be optional, and set the
`$update_from` property based on the database value if it isn't passed.
However, we were using the passed value in the version comparison,
rather than that property, so when assembling the list of updates to run
the database version wasn't actually being taken into account.

See #681