Update user ranks table schema
- Adds the `rank_group`, `blog_id`, and `site_id` columns.
- Adds a unique constraint on `user_id`, `blog_id`, `site_id`,
  `rank_group`.
- Removes `DEFAULT 0` from the `user_id` and `rank_id` columns.

When WordPoints is updated on existing sites, we:

- Drop the defaults.
- Add the new columns.
- Fill in the values of the new columns for existing rows.
- Find and process all duplicate rows, if there are any.
- Add the unique constraint once the duplicates have been dealt with.

When dealing with any duplicates, we:

- Save all of the duplicate rows in an option, for reference.
- Delete all of the affected rows.
- Regenerate the ranks for all of the affected users and rank groups.
- Clear the caches for the affected ranks. (The caches for the ranks and
  users will be cleared automatically as the user ranks are updated.)

Usually there should be no duplicates, but we had to handle them, just
in case, so that we could add the unique constraint.

On multisite regenerating the ranks requires that we switch to the site
for each rank. We utilize the entity context switcher here for better
network switching support on multi-network installs.

Tests are included of course, and the tests for the update to 2.0.0 had
to be updated so that they only ran the update for 2.0.0, and not for
2.4.0. Otherwise, the table updates would run, even though the tables
would not need to be updated, having been created from the new schema to
start with, and this would result in errors. Instead of adding
additional overhead to the test by dropping the tables and creating them
with the old schema, we just skip running the table update tests. This
lead to us fixing several bugs in the update handler, see prior commits.

Fixes #680