Avoid a JOIN when querying for all users with rank
Now that the rank group column is present in the user ranks table, we
can avoid the join.

The tests for the update to the table have to be modified so that they
don't call this function while the old schema is in use, or else the
column won't exist. The function would be called when adding a new rank,
so we now do that before reverting the tables to the old state so that
we can update them.

See #680