Support updating user ranks in bulk
Introduces a new function to update user ranks in bulk, and uses it when
we are transitioning users ranks in the rank group class.

This should improve performance when creating new ranks, since a few
large queries are performed instead of a bunch of little ones.

The tests are also updated to reflect the fact that we no longer have to
run a database query to get the user's current rank at this point. The
single user rank update function had to get the user's old rank, but now
it is passed into the bulk function directly.

See #642