When getting users of base rank limit to current site
When we are querying for all users that have a particular rank, we have
to handle base ranks specially. Users don't have to be assigned base
ranks explicitly; if the user has no rank specified in the database,
then it is assumed that they have the base rank. For this reason, we
have to not only check the user ranks database but also perform an anti
join on the users table to get any users that have no rank. However,
on multisite this ought to be limited to only users who are members of
the current site, since ranks are per-site. But we were including all
users. We've fixed this by letting `WP_User_Query` generate that part
of the query for us, since it automatically limits itself to members of
the current site on multisite.

Fixes #683