Improve memory consumption of refresh user ranks function
Instead of pulling up all users for the previous rank at the same time
as the users of this rank, wait until we actually need them. Duh.

We're going to make more changes than this, but we're doing it
iteratively to show off how ~~bad it was before~~ much better we're
making it.

See #689