PHPCS: Use closing comment style just before closing PHP tag
Otherwise, PHPCS gives an error like:

```
 1 | ERROR | [x] Expected 1 space before closing PHP tag; 0 found
   |       |     (Squiz.PHP.EmbeddedPhp.SpacingBeforeClose)
```

There technically isn't a space before the tag, because it is counted as
part of the comment when using `//` comment style.

I thought about opening a bug for this upstream, but it seems unlikely
to be fixed, as the `Squiz` standard that this sniff is from disallows
comments of this sort entirely:

```
 1 | ERROR | [x] Comments may not appear after statements
   |       |     (Squiz.Commenting.PostStatementComment.Found)
```

So instead, I've decided to work around it by changing the comment style
used in these cases.