phpunit: Avoid infinite loops in extension tests
This would result from the request for the license object trying to
check whether the license was free or not, and not having that info
because we neglected to set it. This would result in it requesting the
info, which would cause the code to be triggered again, because it
would use the version info check to get that information. (The check for
license status performs a different request, and doesn't have this issue
because the response doesn't include the `'new_version'` key.

See #663