Copy post term entity code from hooks API exploration extension
The hooks API extension was created to explore the creation of the new
hooks API. The basic API was merged into core in #321, but the term
entities weren't deemed ready yet at that time. We are now copying the
code relating to them over, which we will iterate on a bit more once it
is in core.

Includes:

- Function to get taxonomies to auto-integrate with.
- Function to get taxonomies to register entities for.
- Function to register entities for a taxonomy.
- Registers entities for taxonomies when the entities app is initialized
  and also hooks to the taxonomy registration function to register
  entities for taxonomies that are registered late.
- Function to register taxonomy entities for a post type.
- Registers term entities relationships for post types when the post
  type entities are registered.

Still needed:

- Tests for the entity classes.
- Term name entity attribute.

See #461