Revamp installables API
The refactoring of the un/installer API also requires the installables
API which was built around it to be refactored accordingly.

The new version of the API almost entirely replaces the old version,
which is now deprecated. The only part that is really retained is the
function to get an un/installer object, since the class name can only be
determined once (it is returned when including the file, which can only
be done once because it defines a class), but in theory may be needed
more than once in a single run.

The new API is designed with high performance in mind, because the check
for whether any updates for any installables need to be run happens on
every single load. This meant that any loading of the installable
classes, or even checking the filesystem to determine whether to load
them (or whether perhaps on the other hand we were dealing with a legacy
un/installer instead), should occur only once we determined that we
might actually need to run the update for the installable in question.

This meant that for one, we needed to be able to specify a loader
function for the installable, rather than having to pass the name of the
class itself. This is especially needed for extensions.

It also meant, that in the interest of performance, we didn't want both
the old and new APIs to be running this check, only the newer, less
intensive version. So this meant that we needed to use a shim to allow
the old un/installers to be run by the new API. This shim installable
could then be returned by the loader function for legacy extensions (or
other installables).

This shim should not be mistaken for a full-fledged installable; it only
provides the shimming code needed for running updates and new site
installs of an un/installer.

This is natural, as in the interest of simplicity we decided that it is
no longer really necessary for install and uninstall to be handled by a
central installables API.

Other than that though, the new installables app provides the same
features as the old static class, with the following differences:

- The update checker now runs the install routine for any entity that
  hasn't been installed, whereas before this was only done for core.
- Admin notices for skipped per-site install/update has been moved to a
  procedural function, since it only relied on the installables registry
  to check if an installable was still actually network active, which
  we've deemed unnecessary. The wording of the notices has also been
  tweaked slightly.

The performance-related improvements actually have minimal affects on
actual operation, and consist of storing a list of versions last checked
in a more comparable format, and ensuring that we only load the
installables that need to be updated JIT.

The only place that we still use the old installables object, other than
for keeping track of the un/installer classes (which we can now do
without having to register them with the `register()` method), is to
call `register()` for any components that still define an un/installer.
This really should be none, because we're going to update core's
components, and there probably shouldn't be any components outside of
core. So this is an edge-case that we're fine calling a deprecated
method for.

We have not yet updated core and the components to use new installable
classes instead of un/installers; that change will be forthcoming.

Some unit tests that we'd have otherwise pursued have been forgone,
simply because the static nature of the old installables and modules
classes are not amenable to unit testing.

See #404